PromptedActions file refactor
